Refactor the Main model structure
A union type for the top-level model proved to be problematic.
It was forcing us to handle the case when the model was not
"bootstrapped" every time we needed to handle page changes, which
was painful. Truely, page data and "app state" data should be
decoupled -- for example, it's feasible that the page state could
be loaded from a cache, while the app state needs to be fetched
from the server.