Remove CSRF requirement from /api/tokens and update token on Phx socket
We don't technically need the CSRF token requirement for the tokens
endpoint, since this is not actually doing anything on the server
other than generating a new stateless JWT and returning it.

We also split the createSocket function into two:

- createPhoenixSocket
- createAbsintheSocket

This way we can hold a reference to the phoenix socket and mutate
it when we have a new token.