Add `if not exists` to creation of extension
I didn't have superuser privelages for my `postgres` user, so it was
easiest for me to log into my dbs and create that extension manually. I
was then getting errors that `citext` already existed, so I figured it
might be easier if that didn't throw an error in the event the
extension is already enabled.