Make all tests async
There were a lot of tests that could have been running async that
weren't tagged as such, so I went ahead and added that tag. On my
machine this decreased test run time from ~4.4 seconds to ~3.2 seconds.

I know, not a huge real-world difference now, but later on when this
test suite gets much bigger, that 25% will probably make a more
significant difference!