Add check for formatting to CI
Keeping things formatted is _way_ easier if you do it starting at the
beginning of a project, and hooking it into CI I've found to be very
helpful. `mix format --check-formatted` will use your `.formatter.exs`
settings to determine if any changes would be made by running `mix
format`, and if any changes would be made, exits with a `1` and fails
the build.