Add dialyzer checks to CI and fix warnings
This adds dialyzer as a check in CI, and fixes the existing dialyzer
warnings. Dialyzer can be a helpful tool, and it can also be really
annoying at times. I will say that in this process I found and fixed a
couple bugs, and I've seen it fix a few other bugs in the past for me
as well.

The biggest downside of dialyzer is that it adds a good chunk of time
to your CI runs, and the actual errors are really obtuse if you haven't
run into them before. The upside is it encourages adding typespecs,
which are really helpful documentation for new contributors (and users
if we're talking about a library, but we're not).

If you want to add it, awesome. If not, that makes sense, too.