Significantly change circleci config
This gets rid of the `build_elixir` job and combines it into the
`test_elixir` job so we can have a single cache key there. There was
actually a good amount of duplication between the `build_elixir` and
`test_elixir` jobs, so this is kind of a deduplication as well.