Add new Globals module to allow requests to mutate repo and session
Before, all pages were setup to be able to send back a new Session
record from the update function (since requests may get performed
that yield an updated session). Now we are encountering the
scenario where the request needs to yield and updated Repo and/or
Session.

This change introduces a Global module, which just contains a
Session and Repo, and has the Post page's update function receive
and return a Globals record instead of a Session.